Add SettingOptionsValue type to settings
Adds a new Settings class "SettingOptionsValue" which allows the user to set
any value in the options list instead of only strings. The new class accepts a list
of dicts with the keys "name" and value", where "name" is the text to show in
the list, and "value" is the value to set when an entry is selected.

It is based on the code for SettingOptions.